requirements.txt
aiohttp
beautifulsoup4
sqlalchemy
psycopg2-binary
python-dotenv
apscheduler
